+++ modules/interrogate/vqa.py
@@ -9,7 +9,7 @@ from modules import shared, devices, errors
 processor = None
 model = None
 loaded: str = None
-MODELS = {
+vlm_models = {
     "MS Florence 2 Base": "microsoft/Florence-2-base", # 0.5GB
     "MS Florence 2 Large": "microsoft/Florence-2-large", # 1.5GB
     "MiaoshouAI PromptGen 1.5 Base": "MiaoshouAI/Florence-2-base-PromptGen-v1.5@c06a5f02cc6071a5d65ee5d294cf3732d3097540", # 1.1GB
@@ -27,17 +27,31 @@ MODELS = {
     "ViLT Base": "dandelin/vilt-b32-finetuned-vqa", # 0.5GB
     "Pix Textcaps": "google/pix2struct-textcaps-base", # 1.1GB
 }
+vlm_prompts = [
+    '<CAPTION>',
+    '<DETAILED_CAPTION>',
+    '<MORE_DETAILED_CAPTION>',
+    '<CAPTION_TO_PHRASE_GROUNDING>',
+    '<OD>',
+    '<DENSE_REGION_CAPTION>',
+    '<REGION_PROPOSAL>',
+    '<OCR>',
+    '<OCR_WITH_REGION>',
+    '<ANALYZE>',
+    '<GENERATE_TAGS>',
+    '<MIXED_CAPTION>',
+    '<MIXED_CAPTION_PLUS>',
+]
 
 
 def git(question: str, image: Image.Image, repo: str = None):
     global processor, model, loaded # pylint: disable=global-statement
     if model is None or loaded != repo:
+        shared.log.debug(f'Interrogate load: vlm="{repo}"')
         model = transformers.GitForCausalLM.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         processor = transformers.GitProcessor.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         loaded = repo
     model.to(devices.device, devices.dtype)
-    shared.log.debug(f'VQA: class={model.__class__.__name__} processor={processor.__class__} model={repo}')
-
     pixel_values = processor(images=image, return_tensors="pt").pixel_values
     git_dict = {}
     git_dict['pixel_values'] = pixel_values.to(devices.device, devices.dtype)
@@ -49,14 +63,13 @@ def git(question: str, image: Image.Image, repo: str = None):
     with devices.inference_context():
         generated_ids = model.generate(**git_dict)
     response = processor.batch_decode(generated_ids, skip_special_tokens=True)[0]
-
-    shared.log.debug(f'VQA: response={response}')
     return response
 
 
 def blip(question: str, image: Image.Image, repo: str = None):
     global processor, model, loaded # pylint: disable=global-statement
     if model is None or loaded != repo:
+        shared.log.debug(f'Interrogate load: vlm="{repo}"')
         model = transformers.BlipForQuestionAnswering.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         processor = transformers.BlipProcessor.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         loaded = repo
@@ -66,21 +79,17 @@ def blip(question: str, image: Image.Image, repo: str = None):
     with devices.inference_context():
         outputs = model.generate(**inputs)
     response = processor.decode(outputs[0], skip_special_tokens=True)
-
-    model.to(devices.cpu)
-    shared.log.debug(f'VQA: response={response}')
     return response
 
 
 def vilt(question: str, image: Image.Image, repo: str = None):
     global processor, model, loaded # pylint: disable=global-statement
     if model is None or loaded != repo:
+        shared.log.debug(f'Interrogate load: vlm="{repo}"')
         model = transformers.ViltForQuestionAnswering.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         processor = transformers.ViltProcessor.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         loaded = repo
     model.to(devices.device)
-    shared.log.debug(f'VQA: class={model.__class__.__name__} processor={processor.__class__} model={repo}')
-
     inputs = processor(image, question, return_tensors="pt")
     inputs = inputs.to(devices.device)
     with devices.inference_context():
@@ -88,20 +97,17 @@ def vilt(question: str, image: Image.Image, repo: str = None):
     logits = outputs.logits
     idx = logits.argmax(-1).item()
     response = model.config.id2label[idx]
-
-    shared.log.debug(f'VQA: response={response}')
     return response
 
 
 def pix(question: str, image: Image.Image, repo: str = None):
     global processor, model, loaded # pylint: disable=global-statement
     if model is None or loaded != repo:
+        shared.log.debug(f'Interrogate load: vlm="{repo}"')
         model = transformers.Pix2StructForConditionalGeneration.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         processor = transformers.Pix2StructProcessor.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         loaded = repo
     model.to(devices.device)
-    shared.log.debug(f'VQA: class={model.__class__.__name__} processor={processor.__class__} model={repo}')
-
     if len(question) > 0:
         inputs = processor(images=image, text=question, return_tensors="pt").to(devices.device)
     else:
@@ -109,14 +115,13 @@ def pix(question: str, image: Image.Image, repo: str = None):
     with devices.inference_context():
         outputs = model.generate(**inputs)
     response = processor.decode(outputs[0], skip_special_tokens=True)
-
-    shared.log.debug(f'VQA: response={response}')
     return response
 
 
 def moondream(question: str, image: Image.Image, repo: str = None):
     global processor, model, loaded # pylint: disable=global-statement
     if model is None or loaded != repo:
+        shared.log.debug(f'Interrogate load: vlm="{repo}"')
         model = transformers.AutoModelForCausalLM.from_pretrained(
             repo,
             revision="2024-08-26",
@@ -127,15 +132,12 @@ def moondream(question: str, image: Image.Image, repo: str = None):
         loaded = repo
         model.eval()
     model.to(devices.device, devices.dtype)
-    shared.log.debug(f'VQA: class={model.__class__.__name__} processor={processor.__class__} model={repo}')
-
     if len(question) < 2:
         question = "Describe the image."
+    question = question.replace('<', '').replace('>', '')
     encoded = model.encode_image(image)
     with devices.inference_context():
         response = model.answer_question(encoded, question, processor)
-
-    shared.log.debug(f'VQA: response="{response}"')
     return response
 
 
@@ -148,6 +150,7 @@ def florence(question: str, image: Image.Image, repo: str = None, revision: str 
             R.remove("flash_attn") # flash_attn is optional
         return R
     if model is None or loaded != repo:
+        shared.log.debug(f'Interrogate load: vlm="{repo}"')
         transformers.dynamic_module_utils.get_imports = get_imports
         model = transformers.AutoModelForCausalLM.from_pretrained(repo, trust_remote_code=True, revision=revision, cache_dir=shared.opts.hfcache_dir)
         processor = transformers.AutoProcessor.from_pretrained(repo, trust_remote_code=True, revision=revision, cache_dir=shared.opts.hfcache_dir)
@@ -155,8 +158,6 @@ def florence(question: str, image: Image.Image, repo: str = None, revision: str 
         loaded = repo
         model.eval()
     model.to(devices.device, devices.dtype)
-    shared.log.debug(f'VQA: class={model.__class__.__name__} processor={processor.__class__} model={repo}')
-
     if question.startswith('<'):
         task = question.split('>', 1)[0] + '>'
     else:
@@ -169,13 +170,12 @@ def florence(question: str, image: Image.Image, repo: str = None, revision: str 
         generated_ids = model.generate(
             input_ids=input_ids,
             pixel_values=pixel_values,
-            max_new_tokens=1024,
-            num_beams=3,
+            max_new_tokens=shared.opts.interrogate_vlm_max_length,
+            num_beams=shared.opts.interrogate_vlm_num_beams,
             do_sample=False
         )
         generated_text = processor.batch_decode(generated_ids, skip_special_tokens=False)[0]
         response = processor.post_process_generation(generated_text, task="task", image_size=(image.width, image.height))
-
     if 'task' in response:
         response = response['task']
     if 'answer' in response:
@@ -183,44 +183,48 @@ def florence(question: str, image: Image.Image, repo: str = None, revision: str 
     if isinstance(response, dict):
         response = json.dumps(response)
     response = response.replace('\n', '').replace('\r', '').replace('\t', '').strip()
-    shared.log.debug(f'VQA: task={task} response="{response}"')
     return response
 
 
-def interrogate(vqa_question, vqa_image, vqa_model_req):
+def interrogate(question, image, model_name):
+    if isinstance(image, list):
+        image = image[0] if len(image) > 0 else None
+    if isinstance(image, dict) and 'name' in image:
+        image = Image.open(image['name'])
+    if image is None:
+        return ''
     try:
-        vqa_model = MODELS.get(vqa_model_req, None)
+        vqa_model = vlm_models.get(model_name, None)
         revision = None
         if '@' in vqa_model:
             vqa_model, revision = vqa_model.split('@')
-        shared.log.debug(f'VQA: model="{vqa_model}" question="{vqa_question}" image={vqa_image}')
-        if vqa_image is None:
+        if image is None:
             answer = 'no image provided'
             return answer
-        if vqa_model_req is None:
+        if model_name is None:
             answer = 'no model selected'
             return answer
         if vqa_model is None:
-            answer = f'unknown: model={vqa_model_req} available={MODELS.keys()}'
+            answer = f'unknown: model={model_name} available={vlm_models.keys()}'
             return answer
         if 'git' in vqa_model.lower():
-            answer = git(vqa_question, vqa_image, vqa_model)
+            answer = git(question, image, vqa_model)
         elif 'vilt' in vqa_model.lower():
-            answer = vilt(vqa_question, vqa_image, vqa_model)
+            answer = vilt(question, image, vqa_model)
         elif 'blip' in vqa_model.lower():
-            answer = blip(vqa_question, vqa_image, vqa_model)
+            answer = blip(question, image, vqa_model)
         elif 'pix' in vqa_model.lower():
-            answer = pix(vqa_question, vqa_image, vqa_model)
+            answer = pix(question, image, vqa_model)
         elif 'moondream2' in vqa_model.lower():
-            answer = moondream(vqa_question, vqa_image, vqa_model)
+            answer = moondream(question, image, vqa_model)
         elif 'florence' in vqa_model.lower():
-            answer = florence(vqa_question, vqa_image, vqa_model, revision)
+            answer = florence(question, image, vqa_model, revision)
         else:
             answer = 'unknown model'
     except Exception as e:
         errors.display(e, 'VQA')
         answer = 'error'
-    if model is not None:
+    if shared.opts.interrogate_offload and model is not None:
         model.to(devices.cpu)
     devices.torch_gc()
     return answer