+++ modules/ui_sections.py
@@ -91,9 +91,11 @@ def create_resolution_inputs(tab):
     return width, height
 
 
-def create_interrogate_button(tab: str, inputs: list, outputs: str):
+def create_interrogate_button(tab: str, inputs: list = None, outputs: str = None):
     button_interrogate = gr.Button(ui_symbols.interrogate, elem_id=f"{tab}_interrogate", elem_classes=['interrogate'])
-    button_interrogate.click(fn=interrogate.interrogate, inputs=inputs, outputs=[outputs])
+    if inputs is not None and outputs is not None:
+        button_interrogate.click(fn=interrogate.interrogate, inputs=inputs, outputs=[outputs])
+    return button_interrogate
 
 
 def create_interrogate_buttons(tab): # legacy function